requirements.txt
flask
torch
torchvision
ultralytics
opencv-python-headless
segment-anything
numpy
Pillow
gunicorn